xmlUtils.h fixed ForceDirected without Warning CmdLineParser made two members const. MessageBox::Show implemented as alias for MessageBox::ShowDialog test: MessageBox added for testing.
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@593 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ test/testdimred.cpp
@@ -69,7 +69,7 @@ void TestFastMap(void) {
     vislib::math::FastMap<myObject, float, OUTDIMS> fm(source, destination);
 
     for (SIZE_T i = 0; i < source.Count(); i++) {
-        printf("item %u: (", i);
+        printf("item %u: (", static_cast<unsigned int>(i));
         for (unsigned int j = 0; j < OUTDIMS; j++) {
             printf("%f", destination[i][j]);
             if (j != OUTDIMS - 1) {
@@ -141,7 +141,7 @@ void TestForceDirected(void) {
     for (SIZE_T loop = 0; loop < 20; loop++) {
         fm.SingleStep();
         for (SIZE_T i = 0; i < source.Count(); i++) {
-            printf("item %u: (", i);
+            printf("item %u: (", static_cast<unsigned int>(i));
             for (unsigned int j = 0; j < 1; j++) {
                 printf("%f", destination[i][j]);
                 if (j != 1 - 1) {
